requirements.txt
asgiref==3.3.1
Django==3.1.6
django-filer==2.0.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==3.0.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
easy-thumbnails==2.7.1
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
Unidecode==1.1.2
